fix: Inline expressions do not have access to classes loaded in Recaf's workspace
