+++ src/main/java/me/coley/recaf/compiler/JavacTargetVersion.java
@@ -11,7 +11,7 @@ import java.lang.reflect.Field;
  *
  * @author Matt
  */
-public enum TargetVersion {
+public enum JavacTargetVersion {
 	V4("1.4"), V5("1.5"), V6("1.6"), V7("1.7"), V8("1.8"), V9("9"), V10("10"), V11("11"), V12("12"), V13("13");
 
 	/**
@@ -19,7 +19,7 @@ public enum TargetVersion {
 	 */
 	private final String opt;
 
-	TargetVersion(String opt) {
+	JavacTargetVersion(String opt) {
 		this.opt = opt;
 	}
 
@@ -36,7 +36,7 @@ public enum TargetVersion {
 	 *
 	 * @return version from class file major version.
 	 */
-	public static TargetVersion fromClassMajor(int version) {
+	public static JavacTargetVersion fromClassMajor(int version) {
 		switch(version) {
 			case Opcodes.V1_4:
 				return V4;
@@ -66,7 +66,7 @@ public enum TargetVersion {
 	/**
 	 * @return Minimum version supported by Javac.
 	 */
-	public static TargetVersion getMinJavacSupport() {
+	public static JavacTargetVersion getMinJavacSupport() {
 		try {
 			Class<?> cls = Class.forName("com.sun.tools.javac.jvm.Target");
 			Field min = cls.getDeclaredField("MIN");
@@ -82,7 +82,7 @@ public enum TargetVersion {
 	/**
 	 * @return Maximum version supported by Javac.
 	 */
-	public static TargetVersion getMaxJavacSupport() {
+	public static JavacTargetVersion getMaxJavacSupport() {
 		try {
 			return fromClassMajor(VMUtil.getVmVersion());
 		} catch (Exception ex) {
