+++ src/main/java/me/coley/recaf/compiler/JavassistCompilationResult.java
@@ -1,0 +1,39 @@
+package me.coley.recaf.compiler;
+
+import javassist.bytecode.Bytecode;
+import javassist.compiler.SymbolTable;
+
+/**
+ * Compilation results.
+ *
+ * @author Matt
+ */
+public class JavassistCompilationResult {
+	private final Bytecode bytecode;
+	private final SymbolTable symbols;
+
+	/**
+	 * @param bytecode
+	 * 		Generated bytecode.
+	 * @param symbols
+	 * 		Generated symbols, if any.
+	 */
+	public JavassistCompilationResult(Bytecode bytecode, SymbolTable symbols) {
+		this.bytecode = bytecode;
+		this.symbols = symbols;
+	}
+
+	/**
+	 * @return Generated bytecode.
+	 */
+	public Bytecode getBytecode() {
+		return bytecode;
+	}
+
+	/**
+	 * @return Generated symbols, if any.
+	 */
+	public SymbolTable getSymbols() {
+		return symbols;
+	}
+}
